Added the files required for the game to function properly.
The files included here are all of the files that were used to complete the project. The project itself was submitted as a zipped file. It is being included here as well for future reference.

The project itself was completed by writing all of the code by hand. The template supplied by Udacity was not used because I wanted to learn how to create a fully functional game from scratch. The animation of the cards was done by shrinking there width and then expanding again to their full width. Using the setTimeout and setInterval methods in JavaScript was quite challenging since everything has to wait in line before the next timer can fire. I added a busy flag to prevent other timed events from walking on each other. Probably not the best approach, but it did get the job done.

I later found out that the animation could have been accomplished using css. I'll post that version shortly.